% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Density function for the triangular distribution}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{Vector of quantiles}

\item{min}{Lower limit of the distribution}

\item{max}{Upper limit of the distribution}

\item{mode}{Mode of the distribution}
}
\value{
Vector of densities
}
\description{
Density function for the triangular distribution
}
