% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function for the triangular distribution}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{Vector of probabilities}

\item{min}{Lower limit of the distribution}

\item{max}{Upper limit of the distribution}

\item{mode}{Mode of the distribution}
}
\value{
Vector of quantiles
}
\description{
Quantile function for the triangular distribution
}
