% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation for the triangular distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{Number of observations}

\item{min}{Lower limit of the distribution}

\item{max}{Upper limit of the distribution}

\item{mode}{Mode of the distribution}
}
\value{
Vector of random values
}
\description{
Random generation for the triangular distribution
}
